requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.7
django-heroku==0.3.1
django-import-export==3.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.5
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
